FIX: mycourses view now displays courses properly based on whether you are a teacher or a student.
When students logged in, they could not see the courses they were
enrolled in. Only teachers could see the courses they hade created.

I fixed the view logic so it detects whether the user is a student
or a teacher and call the appropriate functions. In the JS, I switched
using the created_by to build the query url to just using the user id
and managing the details in the view.