FIX: New school creation works properly.
When I created schools a bug would not allow me to save the
coordinates values into the database. The problem came from the
view implementation. My data would not pass the is_valid method.

I implemented a way for the coordinates to be converted to Points
manually so it would pass the is_valid method and properly save to the
db.