FIX: Buttons are displayed properly following certain conditions in the course view.
The enroll button used to be selectable event when the current
user was already enrolled. Also, teacher students could press the
'cancel' course button even if this was not their course.

The logic has been changed so now only non-enrolled students can press
the enroll button, and only the course creator sees the cancel button.